requirements.txt
# google-cloud-aiplatform>=1.50.0
# google-cloud-secret-manager>=2.20.0
# requests>=2.31.0
# cloudpickle>=3.0.0
# python-dateutil
google-cloud-aiplatform[reasoningengine]
google-cloud-secret-manager
icalendar
python-dotenv
python-dateutil
requests
